fix(courses): extract golfcourseapi mapping out of route files (PR 1 follow-up, #414)
Next route handlers may only export GET/POST/etc. — exporting transformCourse /
normalizeSearch from the route files made next build's generated route types
fail (OmitWithTag '{ [x: string]: never }'), which plain tsc only catches once
.next/types is regenerated (so it slipped the first local check, but CI's build
would fail).

Move the pure mapping into @/lib/golfCourseApi; routes import it and export only
GET. Tests import from the lib module. No behavior change.
