+++ src/app/api/golf-courses/search/route.ts
@@ -1,4 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
+import { API_BASE, normalizeSearch, type CourseSearchResult, type RawSearchCourse } from "@/lib/golfCourseApi";
 
 /**
  * GET /api/golf-courses/search?q=Pebble+Beach
@@ -7,7 +8,8 @@ import { NextRequest, NextResponse } from "next/server";
  * Returns a normalized list of courses (the `id` is the golfcourseapi course
  * identifier; use it with /api/golf-courses/[courseId] to fetch full scorecard
  * data). The normalized shape is unchanged from the prior golfapi.io provider
- * so courseService.ts + CoursePicker.tsx need no changes.
+ * so courseService.ts + CoursePicker.tsx need no changes. Mapping lives in
+ * @/lib/golfCourseApi (route files may only export GET/POST/etc.).
  *
  * Cached for 24h via Next's `revalidate` — golf course metadata barely moves.
  *
@@ -20,31 +22,7 @@ import { NextRequest, NextResponse } from "next/server";
  * is the dumb proxy.
  */
 
-interface CourseSearchResult {
-  id: string;
-  name: string;
-  location: string;
-  city: string;
-  state: string;
-  country: string;
-  courseCount: number;
-}
-
-const SEARCH_LIMIT = 10;
 const REVALIDATE_SECONDS = 60 * 60 * 24; // 24h
-const API_BASE = "https://api.golfcourseapi.com";
-
-// golfcourseapi.com search payload — we type only the fields we read.
-interface RawSearchCourse {
-  id?: number | string;
-  club_name?: string;
-  course_name?: string;
-  location?: {
-    city?: string;
-    state?: string;
-    country?: string;
-  };
-}
 
 export async function GET(req: NextRequest) {
   const apiKey = process.env.GOLFCOURSE_API_KEY;
@@ -80,30 +58,4 @@ export async function GET(req: NextRequest) {
     console.error("golfcourseapi search fetch error:", err);
     return NextResponse.json([] satisfies CourseSearchResult[]);
   }
-}
-
-/** Map golfcourseapi search items into the normalized list shape. Exported for tests. */
-export function normalizeSearch(courses: RawSearchCourse[]): CourseSearchResult[] {
-  return courses.slice(0, SEARCH_LIMIT).map((c) => {
-    const city = c.location?.city ?? "";
-    const stateOrCountry = c.location?.state || c.location?.country || "";
-    const location = [city, stateOrCountry].filter(Boolean).join(", ");
-    // golfcourseapi returns individual courses (not clubs); name leads with the
-    // club, with the course appended when it adds information.
-    const club = c.club_name?.trim() ?? "";
-    const course = c.course_name?.trim() ?? "";
-    const name =
-      club && course && club.toLowerCase() !== course.toLowerCase()
-        ? `${club} — ${course}`
-        : club || course || "Unknown course";
-    return {
-      id: String(c.id ?? ""),
-      name,
-      location,
-      city,
-      state: c.location?.state ?? "",
-      country: c.location?.country ?? "",
-      courseCount: 1,
-    };
-  });
 }
